requirements.txt
aiohttp[speedups]==3.10.5
alembic==1.13.3
asyncpg==0.29.0
atlassian-python-api==3.41.14
click==8.1.7
fastapi==0.115.3
gitpython==3.1.43
google-cloud-firestore==2.17.0
gunicorn==23.0.0
Jinja2==3.1.4
openai==1.35.7
orjson==3.10.7
pre-commit==3.7.0
pydantic==2.8.2
PyGithub==2.3.0
PyJWT==2.9.0
pytest==8.3.3
python-dotenv==1.0.1
python-gitlab==4.9.0
pytz==2024.2
requests==2.32.3
sqlalchemy[asyncio]==2.0.35
tiktoken==0.7.0
tree-sitter-python==0.23.2
tree_sitter==0.22.3
types-pytz==2024.2.0.20240913
types-requests==2.32.0.20240712
uvicorn[standard]==0.32.0
